apps/pipeline/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
pydantic-settings==2.11.0
openai==1.108.1
anthropic==0.72.0
google-generativeai==0.8.5
httpx==0.28.1
python-dotenv==1.1.1
pytest==8.4.1
